Extract ConditionalExpressionHolderHelper from TypeSpecifier
processBooleanConditionalTypes, mergeConditionalHolders and
augmentDisjunctionTypes were only used by BooleanAndHandler and
BooleanOrHandler. Move them (plus their private isTrackableExpression and
isUnsplittableCompoundHolderSide helpers) into a new AutowiredService
ConditionalExpressionHolderHelper injected into both handlers.
